Getting rid of the #present? check.
Selenium::WebDriver::Element#find_element already throws an error if a requested element can not be found. But it throws a very informative ElementNotFound error, whereas this #raise was throwing a useless RuntimeError for the exact same situation.

Several methods make assumptions around ElementNotFoundErrors. For example Wait#until assumes (correctly) that users want errors raised from an element not being found to be ignored. When using GenericElements in a wait block, however, a user would have to tell Wait#until to ignore StandardErrors as well, because this #present? check was essentially masking ElementNotFoundErrors as RuntimeErrors.